+++ src/drivers/motor/motor.h
@@ -1,4 +1,6 @@
+#include <stdint.h>
+
 #pragma once
 
-int setSpeeds(int* speeds);
+int setSpeeds(uint16_t* speeds);
 int initMotors(uint8_t* pins);
